requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt==4.0.1
python-multipart
aiofiles
python-jose
psycopg2-binary
python-dotenv
starlette
jose
requests
beautifulsoup4
alembic